Create compound filters
Create compound filter that represets conditional filters joining
together two or more base or compound filters.

Add AndFilter and OrFilter classes as a convenience for easily
specifying the operator to be used (& or | respectively).

Compound filter has methods for building for adding subfilters to the
current conditional and for creating additional compound filters from
itself and other filter object.